Feature/2056 - Standardize CSV import and export
* Remove old csv upload
