--- src/scenes/shared/simpleTools/upload/index.ts
@@ -1,7 +1,0 @@
-import AdvancedCsvUpload from './AdvancedCsvUpload';
-import CsvUpload from './CsvUpload';
-
-export {
-    AdvancedCsvUpload,
-    CsvUpload
-};
