+++ src/scenes/shared/upload/index.ts
@@ -1,0 +1,5 @@
+import AdvancedCsvUpload from './AdvancedCsvUpload';
+
+export {
+    AdvancedCsvUpload
+};
